add onPrev, onNext, and onClose callbacks
Apps using the viewer might want to react to changes in the viewer.

In particular this will allow updating the url
in the files app
when navigating between files.

This way the url can be copied and used
to navigate to one particular image in a collection.

It also can be used to add browser history entries.
For this to work it is important
to first call the callback
and then update the title.
This way the callback can create a new history entry
and the title of that entry will then be updated.

onPrev and onNext callbacks receive the old and the current fileInfo
as arguments.
onClose callback does not receive any arguments as of now.

Signed-off-by: Azul <azul@riseup.net>